Updated load testing tool
