+++ load-testing/input/baklava/addresses.csv
@@ -1,0 +1,10 @@
+0x352828EF008Cf9f63Cc79ec002449E2d8B16370a
+0x6D73DA97b60520eE0bd54D8c5C7A0Dc5f2ba35AC
+0x4500b26a6F52836510bfd013241DF203f03456F8
+0x26A9928B6e312BB4F8eAaeD5152d05CC87F66201
+0x3EF95251b59d22c540c50870bd999baa31dF717a
+0xFF587B59E7f73a1412fA5C0EB3A02b010F3b60AD
+0xa2343A972f85E7931309A36f5d9353BfE1FB3488
+0x6Af7F878Dc47D3cD83507b760aba82Af45C9F902
+0x123E9F7eDfb0aC47837E4c7B7b7236f64A201096
+0xc9E3B8Df33145F00929F99e97aBBC94aA050690f
