+++ load-testing/input/baklava/celo_validators.csv
@@ -1,0 +1,10 @@
+0x050f34537f5b2a00b9b9c752cb8500a3fce3da7d
+0x2a43f97f8bf959e31f69a894ebd80a88572c8553
+0x30d060f129817c4de5fbc1366d53e19f43c8c64f
+0x456f41406b32c45d59e539e4bba3d7898c3584da
+0x621843731fe33418007c06ee48cfd71e0ea828d9
+0xad682035be6ab6f06e478d2bdab0eab6477b460e
+0xae1ec841923811219b98aceb1db297aade2f46f3
+0xb4e92c94a2712e98c020a81868264bde52c188cb
+0xcda518f6b5a797c3ec45d37c65b83e0b0748edca
+0xdd1f519f63423045f526b8c83edc0eb4ba6434a4
