+++ load-testing/input/mainnet/addresses.csv
@@ -1,0 +1,10 @@
+0x5464D19f5561e7888ea0Cc043715f4356D52D4Cf
+0xb4B7346271497A65cC3Ce36603228e1f57eb57B3
+0x75cdec15A7160F886b33e6366C9dcC457216D873
+0x4D417078055B3d59C81838cDE63A5631972Bb212
+0xda63B8633A07b56d215c822ea8F97A7f99d41197
+0xAA29A80E2a9F5d7e6830297507aCbBF2bD673E41
+0x0c8916127B9ee9339ED0b14695672B43A83283b7
+0xc7d1900659B35E4C5583f25ef016D9caA568Fd05
+0x282F95AdA81610be70D7e1e8557F96866404A632
+0x429CF68a19E304F31730699D5E8D2f3091e850a3
