+++ load-testing/input/mainnet/celo_validators.csv
@@ -1,0 +1,5 @@
+0xc97e88e4803775e87701eedf5bdef8f4a9c89876
+0x5c80858a7a1a730ef6e5d1408f4aed1d5512b080
+0xd22eb6784c69d69275e081283f49c9f46c4b86ea
+0x38c01915f5277f5b18821ba55ecd278379b33666
+0x1dec9eae55fe9348bc195a57df03f1cd51fe13d1
